Фикс ошибок
