--- src/main/java/org/example/Transaction.java
@@ -1,28 +1,0 @@
-package org.example;
-
-import java.math.BigDecimal;
-import java.time.LocalDateTime;
-
-public class Transaction {
-    private String id;
-    private BigDecimal amount;
-    private String type;
-    private LocalDateTime date;
-    private BankAccount sourceAccount;
-    private BankAccount targetAccount;
-
-    public Transaction(String id, BigDecimal amount, String type, LocalDateTime date, BankAccount sourceAccount, BankAccount targetAccount) {
-        this.id = id;
-        this.amount = amount;
-        this.type = type;
-        this.date = date;
-        this.sourceAccount = sourceAccount;
-        this.targetAccount = targetAccount;
-    }
-
-    @Override
-    public String toString() {
-        return String.format("%s || %s || %s || %s || %s || %s",
-                id, amount, type, date, sourceAccount.getOwner(), targetAccount.getOwner());
-    }
-}
