+++ src/main/java/org/example/enums/TranscationType.java
@@ -1,0 +1,7 @@
+package org.example.enums;
+
+public enum TranscationType {
+    DEPOSIT,
+    WITHDRAWAL,
+    TRANSFER
+}
