+++ src/main/java/org/example/exceptions/TransferException.java
@@ -1,0 +1,7 @@
+package org.example.exceptions;
+
+public class TransferException extends RuntimeException {
+    public TransferException(String message) {
+        super(message);
+    }
+}
